requirements.txt
fastmcp
python-dotenv
cachetools